refactor(settings): simplify settings panels by merging related sections
Consolidate 14 settings sections into 8 logical groups to reduce
cognitive load and scroll depth:

- PreferencesSection: Merged Language + Accessibility
- LocationTravelSection: Merged HomeLocation + Transport
- DataProtectionSection: Merged SafeMode + SafeValidation
- HelpToursSection: Merged Tour + Help
- AppInfoSection: Merged About + Updates + Experimental

Kept separate: Profile, DataRetention (privacy-critical), Demo (conditional)

Also adds i18n keys for new section titles in all 4 languages.
