+++ web-app/src/features/settings/components/index.ts
@@ -1,14 +1,8 @@
 export { ProfileSection } from "./ProfileSection";
-export { LanguageSection } from "./LanguageSection";
-export { HomeLocationSection } from "./HomeLocationSection";
-export { TransportSection } from "./TransportSection";
+export { PreferencesSection } from "./PreferencesSection";
+export { LocationTravelSection } from "./LocationTravelSection";
+export { DataProtectionSection } from "./DataProtectionSection";
+export { HelpToursSection } from "./HelpToursSection";
 export { DataRetentionSection } from "./DataRetentionSection";
-export { TourSection } from "./TourSection";
-export { HelpSection } from "./HelpSection";
 export { DemoSection } from "./DemoSection";
-export { AccessibilitySection } from "./AccessibilitySection";
-export { SafeModeSection } from "./SafeModeSection";
-export { SafeValidationSection } from "./SafeValidationSection";
-export { UpdateSection } from "./UpdateSection";
-export { AboutSection } from "./AboutSection";
-export { ExperimentalSection } from "./ExperimentalSection";
+export { AppInfoSection } from "./AppInfoSection";
