+++ web-app/src/i18n/locales/fr.ts
@@ -370,6 +370,18 @@ const fr: Translations = {
     profile: "Profil",
     svNumber: "N° SV",
     language: "Langue",
+    preferences: {
+      title: "Préférences",
+    },
+    locationTravel: {
+      title: "Localisation & Déplacement",
+    },
+    dataProtection: {
+      title: "Protection des données",
+    },
+    helpTours: {
+      title: "Aide & Visites guidées",
+    },
     homeLocation: {
       title: "Emplacement domicile",
       description:
