+++ web-app/src/i18n/locales/it.ts
@@ -368,6 +368,18 @@ const it: Translations = {
     profile: "Profilo",
     svNumber: "N. SV",
     language: "Lingua",
+    preferences: {
+      title: "Preferenze",
+    },
+    locationTravel: {
+      title: "Posizione & Viaggio",
+    },
+    dataProtection: {
+      title: "Protezione dati",
+    },
+    helpTours: {
+      title: "Aiuto & Tour guidati",
+    },
     homeLocation: {
       title: "Posizione casa",
       description:
