+++ web-app/src/i18n/types.ts
@@ -245,6 +245,18 @@ export interface Translations {
     profile: string;
     svNumber: string;
     language: string;
+    preferences: {
+      title: string;
+    };
+    locationTravel: {
+      title: string;
+    };
+    dataProtection: {
+      title: string;
+    };
+    helpTours: {
+      title: string;
+    };
     homeLocation: {
       title: string;
       description: string;
